@@ -8,22 +8,21 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^7.2.5",
+        "php": "^7.4",
         "doctrine/dbal": "^2.10",
-        "fideloper/proxy": "^4.2",
+        "fideloper/proxy": "^4.4",
         "fruitcake/laravel-cors": "^1.0",
         "guzzlehttp/guzzle": "^6.3",
-        "laravel/airlock": "^1.0",
-        "laravel/framework": "^7.0",
-        "laravel/sanctum": "^2.3",
-        "laravel/tinker": "^2.0"
+        "laravel/framework": "^8.0",
+        "laravel/sanctum": "^2.6",
+        "laravel/tinker": "^2.4"
     },
     "require-dev": {
         "facade/ignition": "^2.0",
         "fzaninotto/faker": "^1.9.1",
         "mockery/mockery": "^1.3.1",
-        "nunomaduro/collision": "^4.1",
-        "phpunit/phpunit": "^8.5"
+        "nunomaduro/collision": "^5.0",
+        "phpunit/phpunit": "^9.0"
     },
     "config": {
         "optimize-autoloader": true,
